--- a/dora/services/search.py
+++ b/dora/services/search.py
@@ -168,8 +168,11 @@ def _get_di_results(
         )
     ]
 
+    supported_service_kinds = models.ServiceKind.objects.values_list("value", flat=True)
+
     mapped_di_results = [
-        data_inclusion.map_search_result(result) for result in raw_di_results
+        data_inclusion.map_search_result(result, supported_service_kinds)
+        for result in raw_di_results
     ]
 
     # FIXME: exclu les services uniquement en présentiel à plus de MAX_DISTANCE