Exemples :

Recherche "Toulouse" sans thématique et sans besoin :
On passe de **1560** requêtes en **3.7825s** à **24** requêtes en **1.1775s**

Recherche "Toulouse" avec une thématique et deux besoins :
On passe de **208** requêtes en **0.6408s** à **24** requêtes en **0.5969s**